+++ tests/commands/test_package_releases.py
@@ -10,7 +10,7 @@ def test_command(monkeypatch):
     stdin = read('release_minimal1.json', 'rb') + read('release_minimal2.json', 'rb')
 
     with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))), patch('sys.stdout', new_callable=StringIO) as actual:
-        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'release-package'])
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'package-releases'])
         main()
 
     assert actual.getvalue() == read('release-package_minimal1-minimal2.json')
@@ -20,7 +20,7 @@ def test_command_extensions(monkeypatch):
     stdin = read('release_minimal1.json', 'rb') + read('release_minimal2.json', 'rb')
 
     with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))), patch('sys.stdout', new_callable=StringIO) as actual:
-        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'release-package', 'http://example.com/a/extension.json',
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'package-releases', 'http://example.com/a/extension.json',
                                           'http://example.com/b/extension.json'])
         main()